requirements.txt
Django==2.2.*
Pillow==6.2.*
djangorestframework==3.10.*
